Deprecate TestStore.assert
