+++ Sources/ComposableArchitecture/Internal/Deprecations.swift
@@ -1,6 +1,171 @@
 import CasePaths
 import Combine
 import SwiftUI
+import XCTestDynamicOverlay
+
+// NB: Deprecated after 0.29.0:
+
+#if DEBUG
+  extension TestStore where LocalState: Equatable, Action: Equatable {
+    @available(
+      *, deprecated, message: "Use 'TestStore.send' and 'TestStore.receive' directly, instead"
+    )
+    public func assert(
+      _ steps: Step...,
+      file: StaticString = #file,
+      line: UInt = #line
+    ) {
+      assert(steps, file: file, line: line)
+    }
+
+    @available(
+      *, deprecated, message: "Use 'TestStore.send' and 'TestStore.receive' directly, instead"
+    )
+    public func assert(
+      _ steps: [Step],
+      file: StaticString = #file,
+      line: UInt = #line
+    ) {
+
+      func assert(step: Step) {
+        switch step.type {
+        case let .send(action, update):
+          self.send(action, file: step.file, line: step.line, update)
+
+        case let .receive(expectedAction, update):
+          self.receive(expectedAction, file: step.file, line: step.line, update)
+
+        case let .environment(work):
+          if !self.receivedActions.isEmpty {
+            var actions = ""
+            customDump(self.receivedActions.map(\.action), to: &actions)
+            XCTFail(
+              """
+              Must handle \(self.receivedActions.count) received \
+              action\(self.receivedActions.count == 1 ? "" : "s") before performing this work: …
+
+              Unhandled actions: \(actions)
+              """,
+              file: step.file, line: step.line
+            )
+          }
+          do {
+            try work(&self.environment)
+          } catch {
+            XCTFail("Threw error: \(error)", file: step.file, line: step.line)
+          }
+
+        case let .do(work):
+          if !receivedActions.isEmpty {
+            var actions = ""
+            customDump(self.receivedActions.map(\.action), to: &actions)
+            XCTFail(
+              """
+              Must handle \(self.receivedActions.count) received \
+              action\(self.receivedActions.count == 1 ? "" : "s") before performing this work: …
+
+              Unhandled actions: \(actions)
+              """,
+              file: step.file, line: step.line
+            )
+          }
+          do {
+            try work()
+          } catch {
+            XCTFail("Threw error: \(error)", file: step.file, line: step.line)
+          }
+
+        case let .sequence(subSteps):
+          subSteps.forEach(assert(step:))
+        }
+      }
+
+      steps.forEach(assert(step:))
+
+      self.completed()
+    }
+
+    public struct Step {
+      fileprivate let type: StepType
+      fileprivate let file: StaticString
+      fileprivate let line: UInt
+
+      private init(
+        _ type: StepType,
+        file: StaticString = #file,
+        line: UInt = #line
+      ) {
+        self.type = type
+        self.file = file
+        self.line = line
+      }
+
+      @available(*, deprecated, message: "Call 'TestStore.send' directly, instead")
+      public static func send(
+        _ action: LocalAction,
+        file: StaticString = #file,
+        line: UInt = #line,
+        _ update: @escaping (inout LocalState) throws -> Void = { _ in }
+      ) -> Step {
+        Step(.send(action, update), file: file, line: line)
+      }
+
+      @available(*, deprecated, message: "Call 'TestStore.receive' directly, instead")
+      public static func receive(
+        _ action: Action,
+        file: StaticString = #file,
+        line: UInt = #line,
+        _ update: @escaping (inout LocalState) throws -> Void = { _ in }
+      ) -> Step {
+        Step(.receive(action, update), file: file, line: line)
+      }
+
+      @available(*, deprecated, message: "Mutate 'TestStore.environment' directly, instead")
+      public static func environment(
+        file: StaticString = #file,
+        line: UInt = #line,
+        _ update: @escaping (inout Environment) throws -> Void
+      ) -> Step {
+        Step(.environment(update), file: file, line: line)
+      }
+
+      @available(*, deprecated, message: "Perform this work directly in your test, instead")
+      public static func `do`(
+        file: StaticString = #file,
+        line: UInt = #line,
+        _ work: @escaping () throws -> Void
+      ) -> Step {
+        Step(.do(work), file: file, line: line)
+      }
+
+      @available(*, deprecated, message: "Perform this work directly in your test, instead")
+      public static func sequence(
+        _ steps: [Step],
+        file: StaticString = #file,
+        line: UInt = #line
+      ) -> Step {
+        Step(.sequence(steps), file: file, line: line)
+      }
+
+      @available(*, deprecated, message: "Perform this work directly in your test, instead")
+      public static func sequence(
+        _ steps: Step...,
+        file: StaticString = #file,
+        line: UInt = #line
+      ) -> Step {
+        Step(.sequence(steps), file: file, line: line)
+      }
+
+      fileprivate indirect enum StepType {
+        case send(LocalAction, (inout LocalState) throws -> Void)
+        case receive(Action, (inout LocalState) throws -> Void)
+        case environment((inout Environment) throws -> Void)
+        case `do`(() throws -> Void)
+        case sequence([Step])
+      }
+    }
+  }
+#endif
 
 // NB: Deprecated after 0.27.1:
